Remove 'manifest' as a backup log file that can be fetched
Both the status monitor client and server seem ambivalent about the
manifest entity, what it represents and what can be done with it.

The monitor server represents the manifest as a backup flag in its JSON
output, not as a backup log file that can be fetched by the client, but
at the same time it uses the backup manifest internally as a log file
when handling browse requests.

On input, if the server receives a request to fetch the manifest, such
as c:<client>:b:<number>:l:manifest, it treats the request as valid but
then ignores the requested log file and simply returns the result of
c:<client>:b:<number>.

The client with ncurses UI provides an option to display the manifest of
a selected backup. When selected this option does nothing because the
server does not handle it, as described above.

The changes made here are two-fold:
- The server refuses to handle a request such as
  c:<client>:b:<number>:l:manifest and returns a warning instead.
- The ncurses UI client doesn't show the option to display the manifest
  since it is guaranteed not to work.